requirements-dev.txt
coverage==4.5.1
Django>=2.0
djangorestframework>=3.9
pycodestyle>=2.3
